# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marcenaria API",
        "description": "Documentação da API da Marcenaria usando Laravel + Sanctum + Swagger",
        "contact": {
            "email": "suporte@marcenaria.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Lista todos os serviços",
                "operationId": "97de511018def32ca0ccbc800756c538",
                "responses": {
                    "200": {
                        "description": "Lista de serviços"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Cria um novo serviço",
                "operationId": "0634f2b3ac65bd2815f4b6459f0db97c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome do Serviço"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "serviço-teste"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do serviço"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "Ícone"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço criado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Exibe um serviço específico",
                "operationId": "dce5b306a141bdec91eff8a15534d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serviço encontrado"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Atualiza um serviço",
                "operationId": "14e625e6fca9956d1143de1f41a8a228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome Atualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nova descrição"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Serviço atualizado com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Exclui um serviço",
                "operationId": "9f75ff1ee1d79f882f516173057387ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Serviço excluído com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Lista todas as chaves e valores",
                "operationId": "39919ed520029d628c8b91c83fc87e06",
                "responses": {
                    "200": {
                        "description": "Lista de chaves e valores"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Cria chave e valor",
                "operationId": "5387cb89603507d0535590359d12ef1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "Chave Teste"
                                    },
                                    "valor": {
                                        "type": "string",
                                        "example": "valor-teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chave/valor criado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/settings/{id}": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Exibe chave/valor específico",
                "operationId": "680b5d4953236aa1630ecc00acb65579",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da setting",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chave/valor encontrada"
                    },
                    "404": {
                        "description": "Chave/valor não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Atualiza uma chave/valor",
                "operationId": "d70d9cc5381431c301f12c536297d332",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da chave/valor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "Chave Atualizada"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Novo valor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chave/valor atualizada com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Chave/valor não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Exclui uma chave/valor",
                "operationId": "3593850b3fae0c2fae848e0a55eeaaf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da chave/valor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chave/valor excluída com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Chave/valor não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use o token gerado no login como Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}